requirements.txt
# FastAPI 웹 서버
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# WebSocket
websockets>=12.0
websocket-client>=1.6.0

# 데이터 처리
python-dotenv>=1.0.0

# HTTP API 요청
requests>=2.31.0

# 지오메트리 (지오펜싱)
shapely>=2.0.0
